Category Add in Frontend

+++ app/Http/Controllers/Backend/CategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\CategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
 
@@ -97,10 +98,14 @@ class CategoryController extends Controller
     public function destroy(string $id)
     {
         $category  = Category::findOrFail(id: $id);
-
+        $subCategory = SubCategory::where('category_id', $category->id)->count();
+        if($subCategory>0){
+            return response(['status' => 'error', 'message' => 'This item contain sub items for delete this, you have to delete the sub items first']);
+        }
         $category->delete();
 
-        return response(['status' => 'success', 'Deleted Successfully!']);
+        return response()->json(['status' => 'success', 'message' => 'Deleted Successfully!']);
+
     }
 
     public function changeStatus(Request $request)
