+++ app/Http/Controllers/Backend/SubCategoryController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Backend;
 use App\DataTables\SubCategoryDataTable;
 use App\Http\Controllers\Controller;
 use App\Models\Category;
+use App\Models\ChildCategory;
 use App\Models\SubCategory;
 use Illuminate\Http\Request;
 use Str;
@@ -67,10 +68,7 @@ class SubCategoryController extends Controller
     {
         $categories = Category::all();
         $subCategory = SubCategory::findOrFail($id);
-        return view('admin.sub-category.edit', compact('subCategory', 'categories'));
-        $categories = Category::all();
-        $subCategory = SubCategory::findOrFail($id);
-        return view('admin.sub-category.edit', compact('subCategory', 'categories'));
+        return view(' admin.sub-category.edit', compact('subCategory','categories'));
     }
 
     /**
@@ -88,24 +86,7 @@ class SubCategoryController extends Controller
 
         $subcategory->category_id = $request->category;
         $subcategory->name = $request->name;
-        $subcategory->slug = Str::slug($request->name);
-        $subcategory->status = $request->status;
-        $subcategory->save();
-
-        toastr('Updated Successfully', 'success');
-
-        return redirect()->route('admin.sub-category.index');
-        $request->validate([
-            'category'=>['required'],
-            'name' => ['required', 'max:200', 'unique:sub_categories,name,'.$id],
-            'status' =>['required']
-        ]);
-
-        $subcategory  =SubCategory::findOrFail($id);
-
-        $subcategory->category_id = $request->category;
-        $subcategory->name = $request->name;
-        $subcategory->slug = Str::slug($request->name);
+        $subcategory->slug = Str::slug($request->slug);
         $subcategory->status = $request->status;
         $subcategory->save();
 
@@ -120,7 +101,10 @@ class SubCategoryController extends Controller
     public function destroy(string $id)
     {
         $subCategory  = SubCategory::findOrFail(id: $id);
-
+        $childCategory = ChildCategory::where('sub_category_id', $subCategory->id)->count();
+        if($childCategory>0){
+            return response(['status' => 'error', 'message' => 'This item contain child items for delete this, you have to delete the child items first']);
+        }
         $subCategory->delete();
 
         return response(['status' => 'success', 'Deleted Successfully!']);
